Rewrote backend js for new message tracking, works vastly better now
