+++ IPython/frontend/html/ipythonhttp.py
@@ -7,6 +7,7 @@
 
 # Local imports
 import os
+import threading
 from IPython.external.argparse import ArgumentParser
 from IPython.frontend.html.kernelmanager import HttpKernelManager, \
     IPyHttpServer, IPyHttpHandler
@@ -59,30 +60,38 @@ def main():
                         help='use a GUI widget for tab completion')
 
     args = parser.parse_args()
-
+    
+    def defer():
+        import time
+        time.sleep(2)
+        import webbrowser
+        webbrowser.open("http://localhost:8080/notebook")
+    threading.Thread(target=defer).start()
+    
+    
     # Don't let Qt or ZMQ swallow KeyboardInterupts.
     import signal
     signal.signal(signal.SIGINT, signal.SIG_DFL)
-
-    # Create a KernelManager and start a kernel.
-    kernel_manager = HttpKernelManager(xreq_address=(args.ip, args.xreq),
-                                     sub_address=(args.ip, args.sub),
-                                     rep_address=(args.ip, args.rep),
-                                     hb_address=(args.ip, args.hb))
-    if args.ip == LOCALHOST and not args.existing:
-        if args.pure:
-            kernel_manager.start_kernel(ipython=False)
-        elif args.pylab:
-            kernel_manager.start_kernel(pylab=args.pylab)
-        else:
-            kernel_manager.start_kernel()
-    kernel_manager.start_channels()
-    
-    #Start the web server
-    import webbrowser
-    webbrowser.open("http://localhost:8080/notebook")
-    server = IPyHttpServer(("", 8080), IPyHttpHandler)
-    server.serve_forever()
+    try:
+        # Create a KernelManager and start a kernel.
+        kernel_manager = HttpKernelManager(xreq_address=(args.ip, args.xreq),
+                                         sub_address=(args.ip, args.sub),
+                                         rep_address=(args.ip, args.rep),
+                                         hb_address=(args.ip, args.hb))
+        if args.ip == LOCALHOST and not args.existing:
+            if args.pure:
+                kernel_manager.start_kernel(ipython=False)
+            elif args.pylab:
+                kernel_manager.start_kernel(pylab=args.pylab)
+            else:
+                kernel_manager.start_kernel()
+        kernel_manager.start_channels()
+        
+        #Start the web server
+        server = IPyHttpServer(("", 8080), IPyHttpHandler)
+        server.serve_forever()
+    except KeyboardInterrupt:
+        pass
 
 if __name__ == '__main__':
     main()
