+++ IPython/frontend/html/notebook.html
@@ -13,6 +13,7 @@
         <script type="text/javascript" src="js/notebook.js"></script>
         <script type="text/javascript" src="js/communicate.js"></script>
         <script type="text/javascript" src="js/ui.js"></script>
+        <script type="text/javascript" src="js/manager.js"></script>
     </head>
     <body>
         <div id="messages"></div>
